Move Queue into CI Health
