+++ src/app/ci/queue/page.tsx
@@ -1,0 +1,5 @@
+import { QueueDashboard } from "@/components/queue-dashboard";
+
+export default function CiQueuePage() {
+  return <QueueDashboard />;
+}
